Move hashes from rustc_data_structure to rustc_hashes so they can be shared with rust-analyzer
